Check Lock Door can return possible trap
Added a door property so you can possibly find out if a door is trapped or not when checking a lock.
The chance can also be a false statement so player cant be alway 100% sure of what is happening.

Chances are:
Door trapped (75%)
Door not trapped (25%)

Property is on the door so ALL players were given the same result.

It's a really simple modification that serves for a double job:
- add spicy moments in game and let the player make decisions
- be a sort of workaround to the removed under door mirror function of the optiwand

Might add that to AI officers later.